+++ src/models.ts
@@ -4,6 +4,6 @@ export { ActionType } from './models/action-type.enum';
 export { MapLayerProviderOptions } from './models/map-layer-provider-options.enum';
 export { EntitiesDrawerOptions } from './models/entities-drawer-options';
 export { LayerOptions } from './models/layer-options';
-export { ModuleOptions } from './models/module-options';
+export { ModuleConfiguration } from './models/module-options';
 export { SceneMode } from './models/scene-mode.enum';
 export { KeyboardAction } from './models/ac-keyboard-action.enum';