+++ src/models/module-options.ts
@@ -2,6 +2,6 @@
  * The interface defines the options object that can be passed to AngularCesiumModule on initialization.
  */
 
-export interface ModuleOptions {
+export interface ModuleConfiguration {
   fixEntitiesShadows?: boolean;
 }